Making some changes to the view model. braking but we will survive
